Drop Controleur dependency on Main
We require `IHM.afficherInterface` to be a blocking call.

This allows to remove the dependency on Main in the Controleur class:
the controller does not need to call cleanup code located in Main.

Rather, Main is blocked on `Controleur.demarrer` and will continue its
execution (that is, the cleanup code) when the user interface closes.