Remove Controleur#gererDialogue method
The use of the Command Enum in the Controleur class is a remain of the
oberver pattern.

Calling Controleur#gererDialogue with an Enum and no other arguments,
then just dispatch a call to the UI to call the correct method of
Controleur is useless.

We rather specify a correct interface for Controleur and directly do the
work.
The strength of the coupling remains the same, and the code is easier to
follow.